v2 pay button initial structure
