+++ src/components/v2/V2Project/Pay/V2PayButton.tsx
@@ -1,0 +1,5 @@
+import { Button } from 'antd'
+
+export default function V2PayButton() {
+  return <Button>Pay</Button>
+}
